% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AEEIP_test.R
\name{AEEIP_test}
\alias{AEEIP_test}
\title{#' This is the main function. EM algorithm function and change the beta prior to find better results.}
\usage{
AEEIP_test(
  r_t_pure,
  r_t_mix,
  beta,
  alpha_a = NA,
  alpha_b = NA,
  tau_a_vector,
  l,
  prop_vector = seq(0.01, 0.99, by = 0.01),
  max_iteration = 1e+06,
  stop_stepsize = 1e-07
)
}
\arguments{
\item{r_t_pure}{Read counts of Input sample}

\item{r_t_mix}{Read counts of IP sample}

\item{beta}{the hyper parameter of Beta prior}

\item{alpha_a}{the initial alpha_a}

\item{alpha_b}{the initial alpha_b}

\item{tau_a_vector}{the vector of the initial tau_a}

\item{l}{the effective length of site t}

\item{prop_vector}{the vector of the prop of the beta hyper parameter}

\item{max_iteration}{max iteration}

\item{stop_stepsize}{the threshold for stop}
}
\value{
a list with name c("tau_a_bayes","tau_b_bayes","alpha_a_bayes","alpha_b_bayes",
"tau_a_no_bayes","tau_b_no_bayes","alpha_a_no_bayes",
"alpha_b_no_bayes","num_step",
"difference_value_tau","log_likelihood_bayes",
"log_likelihood_no_bayes","tau_a_bayes_vector",
"tau_a_no_bayes_vector","q_matrix_mix_a","q_matrix_mix_b"
,"beta_a","beta_b","lm_res","tau_a_bayes_vector_aeeip","log_likelihood_bayes_vector",
tau_a_no_bayes_vector_aeeip","log_likelihood_no_bayes_vector","lm_tau_vector")
}
\description{
#' This is the main function. EM algorithm function and change the beta prior to find better results.
}
\examples{
set.seed(1)
data <-simulateData(T=1000, hyper_a_1=20,hyper_b_1 =0.2, hyper_a_2=50,hyper_b_2 =0.05,
pure_rate=0.99,mix_rate=0.1,mode=2,replicates=3)

res <- AEEIP_test(r_t_pure=data[[3]][[1]],r_t_mix=data[[4]][[1]],beta=2,alpha_a=NA,
alpha_b=NA,tau_a_vector=seq(0,0.99,by=0.01),l=1,prop_vector=0.5,
max_iteration =1e6,stop_stepsize=1e-7)
res <- AEEIP_test(r_t_pure=data[[3]][[1]],r_t_mix=data[[4]][[1]],beta=100,alpha_a=NA,
alpha_b=NA,tau_a_vector=1-1e-5,l=1,prop_vector=seq(0.01,0.99,by=0.01),
max_iteration =1e6,stop_stepsize=1e-7)
}
