% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulateData}
\alias{simulateData}
\title{Simulate Data function base on uniform or negative binomial distribution}
\usage{
simulateData(
  T = 2000,
  hyper_a_1 = 0,
  hyper_b_1 = 1000,
  hyper_a_2 = 0,
  hyper_b_2 = 1000,
  pure_rate = 0.999,
  mix_rate = 0.24,
  mode = 1,
  replicates = 3
)
}
\arguments{
\item{T}{number of site}

\item{hyper_a_1}{hyper parameter, minimum for uniform, mean for negative binomial for input}

\item{hyper_b_1}{hyper parameter, maximum for uniform, size for negative binomial for input}

\item{hyper_a_2}{hyper parameter, minimum for uniform, mean for negative binomial for IP}

\item{hyper_b_2}{hyper parameter, maximum for uniform, size for negative binomial for IP}

\item{pure_rate}{the rate of pure sample (Input sample) from RNA fragment A}

\item{mix_rate}{the rate of mix sample (IP sample) from RNA fragment A}

\item{mode}{1 means uniform distribution, 2 means negative binomial}

\item{replicates}{number of replicates}
}
\value{
list(input_real_list,IP_real_list,input_list,IP_mix_list)
input_real_list : real input
IP_real_list : real IP
input_list : mix input
IP_mix_list : mix IP
}
\description{
Simulate Data function base on uniform or negative binomial distribution
}
\examples{


set.seed(1)
data <-simulateData(T=6000, hyper_a_1=0,hyper_b_1 =100, hyper_a_2=0,hyper_b_2 =50,
pure_rate=0.99,mix_rate=0.1,mode=1,replicates=3)

set.seed(1)
data <-simulateData(T=6000, hyper_a_1=20,hyper_b_1 =0.2, hyper_a_2=50,hyper_b_2 =0.05,
pure_rate=0.99,mix_rate=0.1,mode=2,replicates=3)
}
